Updated rate limit test to use new auth
